Everything except updating maps and coordinates working
note to self: dont get distracted doing other things before you have finished doing one feature

+++ CreateModels.cpp
@@ -9,7 +9,7 @@
 void CreateModels(TerrainMap& costMap, vector<vector<IModel*>>& modelMap, IMesh* blockMesh, int mapXSize, int mapYSize)
 {
 	vector<IModel*> row;
-	int modelSizeMultiplier = 10.0f;
+	float modelSizeMultiplier = 10.0f;
 
 	for (int i = 0; i < mapYSize; i++)
 	{
